+++ packages/xo-web/src/xo-app/host/tab-network.js
@@ -2,6 +2,7 @@ import _ from 'intl'
 import ActionButton from 'action-button'
 import Component from 'base-component'
 import copy from 'copy-to-clipboard'
+import humanFormat from 'human-format'
 import React from 'react'
 import Icon from 'icon'
 import SingleLineRow from 'single-line-row'
@@ -263,6 +264,12 @@ const PIF_COLUMNS = [
     itemRenderer: pif => pif.mtu,
     name: _('pifMtuLabel'),
     sortCriteria: 'mtu',
+  },
+  {
+    itemRenderer: ({ speed }) =>
+      speed !== undefined && humanFormat(speed * 1e6, { unit: 'b/s' }), // 1e6: convert Mb to b
+    name: _('pifSpeedLabel'),
+    sortCriteria: 'speed',
   },
   {
     itemRenderer: (pif, userData) => (